projects/sentinel-ai-hub/requirements.txt
azure-ai-projects
azure-ai-agents
azure-search-documents
azure-ai-documentintelligence
azure-identity
azure-eventhub
mcp
python-dotenv
pydantic
openai
